tutor course page

+++ resources/views/components/ui/filepool/file-card.blade.php
@@ -27,11 +27,13 @@
         <button wire:click="downloadFile({{ $file->id }})" class="text-gray-600 hover:text-blue-600 text-sm bg-gray-300 rounded-full p-2 m-2    ">
             <svg xmlns="http://www.w3.org/2000/svg" class="w-5 aspect-square " viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" ><path d="M21 15v4a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2v-4"/><polyline points="7 10 12 15 17 10"/><line x1="12" y1="15" x2="12" y2="3"/></svg>
         </button>
-        <button wire:click="editFile({{ $file->id }})" class="text-gray-600 hover:text-blue-600 text-sm bg-gray-300 rounded-full p-2  m-2 ">
-            <svg xmlns="http://www.w3.org/2000/svg"  class="w-5 aspect-square " viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><path d="M17 3a2.828 2.828 0 1 1 4 4L7.5 20.5 2 22l1.5-5.5L17 3z"/></svg>        
-        </button>
-        <button wire:click="deleteFile({{ $file->id }})"  wire:confirm="Bist du dir sicher? Das Löschen ist unwiederruflich." class="text-red-600 text-sm bg-gray-300 rounded-full p-2  m-2 ">
-            <svg xmlns="http://www.w3.org/2000/svg" class="w-5 aspect-square " viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" ><polyline points="3 6 5 6 21 6"/><path d="M19 6v14a2 2 0 0 1-2 2H7a2 2 0 0 1-2-2V6m3 0V4a2 2 0 0 1 2-2h4a2 2 0 0 1 2 2v2"/><line x1="10" y1="11" x2="10" y2="17"/><line x1="14" y1="11" x2="14" y2="17"/></svg>
-        </button>
+        @if($file->is_owned_by_auth_user)
+            <button wire:click="editFile({{ $file->id }})" class="text-gray-600 hover:text-blue-600 text-sm bg-gray-300 rounded-full p-2  m-2 ">
+                <svg xmlns="http://www.w3.org/2000/svg"  class="w-5 aspect-square " viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><path d="M17 3a2.828 2.828 0 1 1 4 4L7.5 20.5 2 22l1.5-5.5L17 3z"/></svg>        
+            </button>
+            <button wire:click="deleteFile({{ $file->id }})"  wire:confirm="Bist du dir sicher? Das Löschen ist unwiederruflich." class="text-red-600 text-sm bg-gray-300 rounded-full p-2  m-2 ">
+                <svg xmlns="http://www.w3.org/2000/svg" class="w-5 aspect-square " viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" ><polyline points="3 6 5 6 21 6"/><path d="M19 6v14a2 2 0 0 1-2 2H7a2 2 0 0 1-2-2V6m3 0V4a2 2 0 0 1 2-2h4a2 2 0 0 1 2 2v2"/><line x1="10" y1="11" x2="10" y2="17"/><line x1="14" y1="11" x2="14" y2="17"/></svg>
+            </button>
+        @endif
     </div>
 </div>
